Share SignalR's HubProtocol implementation in Blazor (#20922)
